Refactor Http2ClientHanlder to delegate upgrade request handing to a separate handler
